reconsitution of class Bintray and Atlas
